# bpo-40255: Implement Immortal Instances - Optimization 3